refactored DigitalTwins and Pnp samples to initialize loggers from Sample.cs files instead of Program.cs
