+++ iot-hub/Samples/service/DigitalTwinClientSamples/TemperatureController/Program.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using CommandLine;
-using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
 
@@ -10,8 +9,6 @@ namespace Microsoft.Azure.Devices.Samples
 {
     public class Program
     {
-        private static ILogger s_logger;
-
         /// <summary>
         /// This sample performs component-level operations on a plug and play compatible device.
         /// </summary>
@@ -32,33 +29,15 @@ namespace Microsoft.Azure.Devices.Samples
                     Environment.Exit(1);
                 });
 
-            s_logger = InitializeConsoleDebugLogger();
             if (!parameters.Validate())
             {
                 throw new ArgumentException("Required parameters are not set. Please recheck required variables by using \"--help\"");
             }
 
-            s_logger.LogDebug("Set up the digital twin client.");
             using DigitalTwinClient digitalTwinClient = DigitalTwinClient.CreateFromConnectionString(parameters.HubConnectionString);
 
-            s_logger.LogDebug("Set up and start the TemperatureController service sample.");
-            var temperatureControllerSample = new TemperatureControllerSample(digitalTwinClient, parameters.DeviceId, s_logger);
+            var temperatureControllerSample = new TemperatureControllerSample(digitalTwinClient, parameters.DeviceId);
             await temperatureControllerSample.RunSampleAsync().ConfigureAwait(false);
-        }
-
-        private static ILogger InitializeConsoleDebugLogger()
-        {
-            using ILoggerFactory loggerFactory = LoggerFactory.Create(builder =>
-            {
-                builder
-                .AddFilter(level => level >= LogLevel.Debug)
-                .AddSimpleConsole(options =>
-                {
-                    options.TimestampFormat = "[MM/dd/yyyy HH:mm:ss]";
-                });
-            });
-
-            return loggerFactory.CreateLogger<TemperatureControllerSample>();
         }
     }
 }
