+++ iot-hub/Samples/service/PnpServiceSamples/Thermostat/Program.cs
@@ -3,7 +3,6 @@
 
 using CommandLine;
 using Microsoft.Azure.Devices.Shared;
-using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System;
 using System.Threading.Tasks;
@@ -12,8 +11,6 @@ namespace Microsoft.Azure.Devices.Samples
 {
     public class Program
     {
-        private static ILogger s_logger;
-
         /// <summary>
         /// This sample performs root-level operations on a plug and play compatible device using the IoT Hub service client
         /// </summary>
@@ -34,34 +31,16 @@ namespace Microsoft.Azure.Devices.Samples
                     Environment.Exit(1);
                 });
 
-            s_logger = InitializeConsoleDebugLogger();
             if (!parameters.Validate())
             {
                 throw new ArgumentException("Required parameters are not set. Please recheck required variables by using \"--help\"");
             }
 
-            s_logger.LogDebug("Set up the IoT Hub service client and registry manager.");
             using ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(parameters.HubConnectionString);
             using RegistryManager registryManager = RegistryManager.CreateFromConnectionString(parameters.HubConnectionString);
 
-            s_logger.LogDebug("Set up and start the Thermostat service sample.");
-            var thermostatSample = new ThermostatSample(serviceClient, registryManager, parameters.DeviceId, s_logger);
+            var thermostatSample = new ThermostatSample(serviceClient, registryManager, parameters.DeviceId);
             await thermostatSample.RunSampleAsync().ConfigureAwait(false);
-        }
-
-        private static ILogger InitializeConsoleDebugLogger()
-        {
-            using ILoggerFactory loggerFactory = LoggerFactory.Create(builder =>
-            {
-                builder
-                .AddFilter(level => level >= LogLevel.Debug)
-                .AddSimpleConsole(options =>
-                {
-                    options.TimestampFormat = "[MM/dd/yyyy HH:mm:ss]";
-                });
-            });
-
-            return loggerFactory.CreateLogger<ThermostatSample>();
         }
     }
 }
