+++ iot-hub/Samples/service/PnpServiceSamples/Thermostat/ThermostatSample.cs
@@ -18,12 +18,12 @@ namespace Microsoft.Azure.Devices.Samples
         private readonly string _deviceId;
         private readonly ILogger _logger;
 
-        public ThermostatSample(ServiceClient serviceClient, RegistryManager registryManager, string deviceId, ILogger logger)
+        public ThermostatSample(ServiceClient serviceClient, RegistryManager registryManager, string deviceId)
         {
             _serviceClient = serviceClient ?? throw new ArgumentNullException(nameof(serviceClient));
             _deviceId = deviceId ?? throw new ArgumentNullException(nameof(deviceId));
             _registryManager = registryManager ?? throw new ArgumentNullException(nameof(registryManager));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _logger = InitializeConsoleDebugLogger();
         }
 
         public async Task RunSampleAsync()
@@ -37,6 +37,21 @@ namespace Microsoft.Azure.Devices.Samples
 
             // Invoke the root-level command getMaxMinReport command on the device twin
             await InvokeGetMaxMinReportCommandAsync();
+        }
+
+        private static ILogger InitializeConsoleDebugLogger()
+        {
+            using ILoggerFactory loggerFactory = LoggerFactory.Create(builder =>
+            {
+                builder
+                .AddFilter(level => level >= LogLevel.Debug)
+                .AddSimpleConsole(options =>
+                {
+                    options.TimestampFormat = "[MM/dd/yyyy HH:mm:ss]";
+                });
+            });
+
+            return loggerFactory.CreateLogger<ThermostatSample>();
         }
 
         private async Task<Twin> GetAndPrintDeviceTwinAsync()
